requirements.txt
discord.py
python-dotenv
uuid
